+++ crates/db/Cargo.toml
@@ -22,5 +22,7 @@ moor-values = { path = "../values" }
 ## Error declaration/ handling
 thiserror.workspace = true
 
-uuid.workspace = true
+strum.workspace = true
 tracing.workspace = true
+uuid.workspace = true
+
