+++ crates/db/src/worldstate_tests.rs
@@ -1,0 +1,948 @@
+//! A set of common tests for any world state implementation.
+
+use crate::worldstate_transaction::WorldStateTransaction;
+use crate::{RelationalTransaction, RelationalWorldStateTransaction, WorldStateTable};
+use moor_values::model::ObjSet;
+use moor_values::model::VerbArgsSpec;
+use moor_values::model::{BinaryType, VerbAttrs};
+use moor_values::model::{CommitResult, WorldStateError};
+use moor_values::model::{HasUuid, Named};
+use moor_values::model::{ObjAttrs, PropFlag, ValSet};
+use moor_values::util::BitEnum;
+use moor_values::var::Objid;
+use moor_values::var::{v_int, v_str};
+use moor_values::NOTHING;
+
+pub fn perform_test_create_object<F, TX>(begin_tx: F)
+where
+    F: Fn() -> RelationalWorldStateTransaction<TX>,
+    TX: RelationalTransaction<WorldStateTable>,
+{
+    let tx = begin_tx();
+    let oid = tx
+        .create_object(
+            None,
+            ObjAttrs::new(NOTHING, NOTHING, NOTHING, BitEnum::new(), "test"),
+        )
+        .unwrap();
+    assert_eq!(oid, Objid(0));
+    assert!(tx.object_valid(oid).unwrap());
+    assert_eq!(tx.get_object_owner(oid).unwrap(), oid);
+    assert_eq!(tx.get_object_parent(oid).unwrap(), NOTHING);
+    assert_eq!(tx.get_object_location(oid).unwrap(), NOTHING);
+    assert_eq!(tx.get_object_name(oid).unwrap(), "test");
+    assert_eq!(tx.commit(), Ok(CommitResult::Success));
+
+    // Verify existence in a new transaction.
+    let tx = begin_tx();
+    assert!(tx.object_valid(oid).unwrap());
+    assert_eq!(tx.get_object_owner(oid).unwrap(), oid);
+}
+
+pub fn perform_test_create_object_fixed_id<F, TX>(begin_tx: F)
+where
+    F: Fn() -> RelationalWorldStateTransaction<TX>,
+    TX: RelationalTransaction<WorldStateTable>,
+{
+    let tx = begin_tx();
+
+    // Force at 1.
+    let oid = tx
+        .create_object(Some(Objid(1)), ObjAttrs::default())
+        .unwrap();
+    assert_eq!(oid, Objid(1));
+    // Now verify the next will be 2.
+    let oid2 = tx.create_object(None, ObjAttrs::default()).unwrap();
+    assert_eq!(oid2, Objid(2));
+    assert_eq!(tx.commit(), Ok(CommitResult::Success));
+}
+
+pub fn perform_test_parent_children<F, TX>(begin_tx: F)
+where
+    F: Fn() -> RelationalWorldStateTransaction<TX>,
+    TX: RelationalTransaction<WorldStateTable>,
+{
+    let tx = begin_tx();
+
+    // Single parent/child relationship.
+    let a = tx
+        .create_object(
+            None,
+            ObjAttrs::new(NOTHING, NOTHING, NOTHING, BitEnum::new(), "test"),
+        )
+        .unwrap();
+
+    let b = tx
+        .create_object(
+            None,
+            ObjAttrs::new(NOTHING, a, NOTHING, BitEnum::new(), "test2"),
+        )
+        .unwrap();
+
+    assert_eq!(tx.get_object_parent(b).unwrap(), a);
+    assert!(tx
+        .get_object_children(a)
+        .unwrap()
+        .is_same(ObjSet::from_items(&[b])));
+
+    assert_eq!(tx.get_object_parent(a).unwrap(), NOTHING);
+    assert_eq!(tx.get_object_children(b).unwrap(), ObjSet::empty());
+
+    // Add a second child
+    let c = tx
+        .create_object(
+            None,
+            ObjAttrs::new(NOTHING, a, NOTHING, BitEnum::new(), "test3"),
+        )
+        .unwrap();
+
+    assert_eq!(tx.get_object_parent(c).unwrap(), a);
+    assert!(tx
+        .get_object_children(a)
+        .unwrap()
+        .is_same(ObjSet::from_items(&[b, c])));
+
+    assert_eq!(tx.get_object_parent(a).unwrap(), NOTHING);
+    assert_eq!(tx.get_object_children(b).unwrap(), ObjSet::empty());
+
+    // Create new obj and reparent one child
+    let d = tx
+        .create_object(
+            None,
+            ObjAttrs::new(NOTHING, NOTHING, NOTHING, BitEnum::new(), "test4"),
+        )
+        .unwrap();
+
+    tx.set_object_parent(b, d).unwrap();
+    assert_eq!(tx.get_object_parent(b).unwrap(), d);
+    assert!(tx
+        .get_object_children(a)
+        .unwrap()
+        .is_same(ObjSet::from_items(&[c])));
+    assert!(tx
+        .get_object_children(d)
+        .unwrap()
+        .is_same(ObjSet::from_items(&[b])));
+    assert_eq!(tx.commit(), Ok(CommitResult::Success));
+}
+
+pub fn perform_test_descendants<F, TX>(begin_tx: F)
+where
+    F: Fn() -> RelationalWorldStateTransaction<TX>,
+    TX: RelationalTransaction<WorldStateTable>,
+{
+    let tx = begin_tx();
+
+    let a = tx
+        .create_object(
+            None,
+            ObjAttrs::new(NOTHING, NOTHING, NOTHING, BitEnum::new(), "test"),
+        )
+        .unwrap();
+    assert_eq!(a, Objid(0));
+
+    let b = tx
+        .create_object(
+            None,
+            ObjAttrs::new(NOTHING, a, NOTHING, BitEnum::new(), "test2"),
+        )
+        .unwrap();
+    assert_eq!(b, Objid(1));
+
+    let c = tx
+        .create_object(
+            None,
+            ObjAttrs::new(NOTHING, a, NOTHING, BitEnum::new(), "test3"),
+        )
+        .unwrap();
+    assert_eq!(c, Objid(2));
+
+    let d = tx
+        .create_object(
+            None,
+            ObjAttrs::new(NOTHING, c, NOTHING, BitEnum::new(), "test4"),
+        )
+        .unwrap();
+    assert_eq!(d, Objid(3));
+
+    assert!(tx
+        .descendants(a)
+        .unwrap()
+        .is_same(ObjSet::from_items(&[b, c, d])));
+    assert_eq!(tx.descendants(b).unwrap(), ObjSet::empty());
+    assert_eq!(tx.descendants(c).unwrap(), ObjSet::from_items(&[d]));
+
+    // Now reparent d to b
+    tx.set_object_parent(d, b).unwrap();
+    assert!(tx
+        .get_object_children(a)
+        .unwrap()
+        .is_same(ObjSet::from_items(&[b, c])));
+    assert_eq!(tx.get_object_children(b).unwrap(), ObjSet::from_items(&[d]));
+    assert_eq!(tx.get_object_children(c).unwrap(), ObjSet::empty());
+    assert!(tx
+        .descendants(a)
+        .unwrap()
+        .is_same(ObjSet::from_items(&[b, c, d])));
+    assert_eq!(tx.descendants(b).unwrap(), ObjSet::from_items(&[d]));
+    assert_eq!(tx.descendants(c).unwrap(), ObjSet::empty());
+    assert_eq!(tx.commit(), Ok(CommitResult::Success));
+}
+
+pub fn perform_test_location_contents<F, TX>(begin_tx: F)
+where
+    F: Fn() -> RelationalWorldStateTransaction<TX>,
+    TX: RelationalTransaction<WorldStateTable>,
+{
+    let tx = begin_tx();
+
+    let a = tx
+        .create_object(
+            None,
+            ObjAttrs::new(NOTHING, NOTHING, NOTHING, BitEnum::new(), "test"),
+        )
+        .unwrap();
+
+    let b = tx
+        .create_object(
+            None,
+            ObjAttrs::new(NOTHING, NOTHING, a, BitEnum::new(), "test2"),
+        )
+        .unwrap();
+
+    assert_eq!(tx.get_object_location(b).unwrap(), a);
+    assert_eq!(tx.get_object_contents(a).unwrap(), ObjSet::from_items(&[b]));
+
+    assert_eq!(tx.get_object_location(a).unwrap(), NOTHING);
+    assert_eq!(tx.get_object_contents(b).unwrap(), ObjSet::empty());
+
+    let c = tx
+        .create_object(
+            None,
+            ObjAttrs::new(NOTHING, NOTHING, NOTHING, BitEnum::new(), "test3"),
+        )
+        .unwrap();
+
+    tx.set_object_location(b, c).unwrap();
+    assert_eq!(tx.get_object_location(b).unwrap(), c);
+    assert_eq!(tx.get_object_contents(a).unwrap(), ObjSet::empty());
+    assert_eq!(tx.get_object_contents(c).unwrap(), ObjSet::from_items(&[b]));
+
+    let d = tx
+        .create_object(
+            None,
+            ObjAttrs::new(NOTHING, NOTHING, NOTHING, BitEnum::new(), "test4"),
+        )
+        .unwrap();
+    tx.set_object_location(d, c).unwrap();
+    assert!(tx
+        .get_object_contents(c)
+        .unwrap()
+        .is_same(ObjSet::from_items(&[b, d])));
+    assert_eq!(tx.get_object_location(d).unwrap(), c);
+
+    tx.set_object_location(a, c).unwrap();
+    assert!(tx
+        .get_object_contents(c)
+        .unwrap()
+        .is_same(ObjSet::from_items(&[b, d, a])));
+    assert_eq!(tx.get_object_location(a).unwrap(), c);
+
+    // Validate recursive move detection.
+    match tx.set_object_location(c, b).err() {
+        Some(WorldStateError::RecursiveMove(_, _)) => {}
+        _ => {
+            panic!("Expected recursive move error");
+        }
+    }
+
+    // Move b one level deeper, and then check recursive move detection again.
+    tx.set_object_location(b, d).unwrap();
+    match tx.set_object_location(c, b).err() {
+        Some(WorldStateError::RecursiveMove(_, _)) => {}
+        _ => {
+            panic!("Expected recursive move error");
+        }
+    }
+
+    // The other way around, d to c should be fine.
+    tx.set_object_location(d, c).unwrap();
+    assert_eq!(tx.commit(), Ok(CommitResult::Success));
+}
+
+/// Test data integrity of object moves between commits.
+pub fn perform_test_object_move_commits<F, TX>(begin_tx: F)
+where
+    F: Fn() -> RelationalWorldStateTransaction<TX>,
+    TX: RelationalTransaction<WorldStateTable>,
+{
+    let tx = begin_tx();
+
+    let a = tx
+        .create_object(
+            None,
+            ObjAttrs::new(NOTHING, NOTHING, NOTHING, BitEnum::new(), "test"),
+        )
+        .unwrap();
+
+    let b = tx
+        .create_object(
+            None,
+            ObjAttrs::new(NOTHING, NOTHING, a, BitEnum::new(), "test2"),
+        )
+        .unwrap();
+
+    let c = tx
+        .create_object(
+            None,
+            ObjAttrs::new(NOTHING, NOTHING, NOTHING, BitEnum::new(), "test3"),
+        )
+        .unwrap();
+
+    tx.set_object_location(b, a).unwrap();
+    tx.set_object_location(c, a).unwrap();
+    assert_eq!(tx.get_object_location(b).unwrap(), a);
+    assert_eq!(tx.get_object_location(c).unwrap(), a);
+    assert!(tx
+        .get_object_contents(a)
+        .unwrap()
+        .is_same(ObjSet::from_items(&[b, c])));
+    assert_eq!(tx.get_object_contents(b).unwrap(), ObjSet::empty());
+    assert_eq!(tx.get_object_contents(c).unwrap(), ObjSet::empty());
+
+    assert_eq!(tx.commit(), Ok(CommitResult::Success));
+
+    let tx = begin_tx();
+    assert_eq!(tx.get_object_location(b).unwrap(), a);
+    assert_eq!(tx.get_object_location(c).unwrap(), a);
+    let contents = tx.get_object_contents(a).expect("Unable to get contents");
+    assert!(
+        contents.is_same(ObjSet::from_items(&[b, c])),
+        "Contents of a are not as expected: {:?} vs {:?}",
+        contents,
+        ObjSet::from_items(&[b, c])
+    );
+    assert_eq!(tx.get_object_contents(b).unwrap(), ObjSet::empty());
+    assert_eq!(tx.get_object_contents(c).unwrap(), ObjSet::empty());
+
+    tx.set_object_location(b, c).unwrap();
+    assert_eq!(tx.get_object_location(b).unwrap(), c);
+    assert_eq!(tx.get_object_location(c).unwrap(), a);
+    assert_eq!(tx.get_object_contents(a).unwrap(), ObjSet::from_items(&[c]));
+    assert_eq!(tx.get_object_contents(b).unwrap(), ObjSet::empty());
+    assert_eq!(tx.get_object_contents(c).unwrap(), ObjSet::from_items(&[b]));
+    assert_eq!(tx.commit(), Ok(CommitResult::Success));
+
+    let tx = begin_tx();
+    assert_eq!(tx.get_object_location(c).unwrap(), a);
+    assert_eq!(tx.get_object_location(b).unwrap(), c);
+    assert_eq!(tx.get_object_contents(a).unwrap(), ObjSet::from_items(&[c]));
+    assert_eq!(tx.get_object_contents(b).unwrap(), ObjSet::empty());
+    assert_eq!(tx.get_object_contents(c).unwrap(), ObjSet::from_items(&[b]));
+}
+
+pub fn perform_test_simple_property<F, TX>(begin_tx: F)
+where
+    F: Fn() -> RelationalWorldStateTransaction<TX>,
+    TX: RelationalTransaction<WorldStateTable>,
+{
+    let tx = begin_tx();
+
+    let oid = tx
+        .create_object(
+            None,
+            ObjAttrs::new(NOTHING, NOTHING, NOTHING, BitEnum::new(), "test"),
+        )
+        .unwrap();
+
+    tx.define_property(
+        oid,
+        oid,
+        "test".into(),
+        NOTHING,
+        BitEnum::new(),
+        Some(v_str("test")),
+    )
+    .unwrap();
+    let (prop, v, perms, is_clear) = tx.resolve_property(oid, "test".into()).unwrap();
+    assert_eq!(prop.name(), "test");
+    assert_eq!(v, v_str("test"));
+    assert_eq!(perms.owner(), NOTHING);
+    assert!(!is_clear);
+    assert_eq!(tx.commit(), Ok(CommitResult::Success));
+}
+
+/// Regression test for updating-verbs failing.
+pub fn perform_test_verb_add_update<F, TX>(begin_tx: F)
+where
+    F: Fn() -> RelationalWorldStateTransaction<TX>,
+    TX: RelationalTransaction<WorldStateTable>,
+{
+    let tx = begin_tx();
+    let oid = tx
+        .create_object(
+            None,
+            ObjAttrs::new(NOTHING, NOTHING, NOTHING, BitEnum::new(), "test"),
+        )
+        .unwrap();
+    tx.add_object_verb(
+        oid,
+        oid,
+        vec!["test".into()],
+        vec![],
+        BinaryType::LambdaMoo18X,
+        BitEnum::new(),
+        VerbArgsSpec::this_none_this(),
+    )
+    .unwrap();
+    // resolve the verb to its vh.
+    let vh = tx.resolve_verb(oid, "test".into(), None).unwrap();
+    assert_eq!(vh.names(), vec!["test"]);
+    // Verify it's actually on the object when we get verbs.
+    let verbs = tx.get_verbs(oid).unwrap();
+    assert_eq!(verbs.len(), 1);
+    assert!(verbs.contains(vh.uuid()));
+    // update the verb using its uuid, renaming it.
+    tx.update_verb(
+        oid,
+        vh.uuid(),
+        VerbAttrs {
+            definer: None,
+            owner: None,
+            names: Some(vec!["test2".into()]),
+            flags: None,
+            args_spec: None,
+            binary_type: None,
+            binary: None,
+        },
+    )
+    .unwrap();
+    // resolve with the new name.
+    let vh = tx.resolve_verb(oid, "test2".into(), None).unwrap();
+    assert_eq!(vh.names(), vec!["test2"]);
+
+    // Now commit, and try to resolve again.
+    assert_eq!(tx.commit(), Ok(CommitResult::Success));
+    let tx = begin_tx();
+    let vh = tx.resolve_verb(oid, "test2".into(), None).unwrap();
+    assert_eq!(vh.names(), vec!["test2"]);
+    assert_eq!(tx.commit(), Ok(CommitResult::Success));
+}
+
+pub fn perform_test_transitive_property_resolution<F, TX>(begin_tx: F)
+where
+    F: Fn() -> RelationalWorldStateTransaction<TX>,
+    TX: RelationalTransaction<WorldStateTable>,
+{
+    let tx = begin_tx();
+
+    let a = tx
+        .create_object(
+            None,
+            ObjAttrs::new(NOTHING, NOTHING, NOTHING, BitEnum::new(), "test"),
+        )
+        .unwrap();
+
+    let b = tx
+        .create_object(
+            None,
+            ObjAttrs::new(NOTHING, a, NOTHING, BitEnum::new(), "test2"),
+        )
+        .unwrap();
+
+    tx.define_property(
+        a,
+        a,
+        "test".into(),
+        NOTHING,
+        BitEnum::new(),
+        Some(v_str("test_value")),
+    )
+    .unwrap();
+    let (prop, v, perms, is_clear) = tx.resolve_property(b, "test".into()).unwrap();
+    assert_eq!(prop.name(), "test");
+    assert_eq!(v, v_str("test_value"));
+    assert_eq!(perms.owner(), NOTHING);
+    assert!(is_clear);
+
+    // Verify we *don't* get this property for an unrelated, unhinged object by reparenting b
+    // to new parent c.  This should remove the defs for a's properties from b.
+    let c = tx
+        .create_object(
+            None,
+            ObjAttrs::new(NOTHING, NOTHING, NOTHING, BitEnum::new(), "test3"),
+        )
+        .unwrap();
+
+    tx.set_object_parent(b, c).unwrap();
+
+    let result = tx.resolve_property(b, "test".into());
+    assert_eq!(
+        result.err().unwrap(),
+        WorldStateError::PropertyNotFound(b, "test".into())
+    );
+    assert_eq!(tx.commit(), Ok(CommitResult::Success));
+}
+
+pub fn perform_test_transitive_property_resolution_clear_property<F, TX>(begin_tx: F)
+where
+    F: Fn() -> RelationalWorldStateTransaction<TX>,
+    TX: RelationalTransaction<WorldStateTable>,
+{
+    let tx = begin_tx();
+
+    let a = tx
+        .create_object(
+            None,
+            ObjAttrs::new(NOTHING, NOTHING, NOTHING, BitEnum::new(), "test"),
+        )
+        .unwrap();
+
+    let b = tx
+        .create_object(
+            None,
+            ObjAttrs::new(NOTHING, a, NOTHING, BitEnum::new(), "test2"),
+        )
+        .unwrap();
+
+    tx.define_property(
+        a,
+        a,
+        "test".into(),
+        NOTHING,
+        BitEnum::new(),
+        Some(v_str("test_value")),
+    )
+    .unwrap();
+    let (prop, v, perms, is_clear) = tx.resolve_property(b, "test".into()).unwrap();
+    assert_eq!(prop.name(), "test");
+    assert_eq!(v, v_str("test_value"));
+    assert_eq!(perms.owner(), NOTHING);
+    assert!(is_clear);
+
+    // Set the property on the child to a new value.
+    tx.set_property(b, prop.uuid(), v_int(666)).unwrap();
+
+    // Verify the new value is present.
+    let (prop, v, perms, is_clear) = tx.resolve_property(b, "test".into()).unwrap();
+    assert_eq!(prop.name(), "test");
+    assert_eq!(v, v_int(666));
+    assert_eq!(perms.owner(), NOTHING);
+    assert!(!is_clear);
+
+    // Now clear, and we should get the old value, but with clear status.
+    tx.clear_property(b, prop.uuid()).unwrap();
+    let (prop, v, perms, is_clear) = tx.resolve_property(b, "test".into()).unwrap();
+    assert_eq!(prop.name(), "test");
+    assert_eq!(v, v_str("test_value"));
+    assert_eq!(perms.owner(), NOTHING);
+    assert!(is_clear);
+
+    // Changing flags or owner should have nothing to do with the clarity of the property value.
+    tx.update_property_info(
+        b,
+        prop.uuid(),
+        Some(b),
+        Some(BitEnum::new_with(PropFlag::Read)),
+        None,
+    )
+    .unwrap();
+    let (prop, v, perms, is_clear) = tx.resolve_property(b, "test".into()).unwrap();
+    assert_eq!(prop.name(), "test");
+    assert_eq!(v, v_str("test_value"));
+    assert_eq!(perms.owner(), b);
+    assert_eq!(perms.flags(), BitEnum::new_with(PropFlag::Read));
+    assert!(is_clear);
+
+    // Setting the value again makes it not clear
+    tx.set_property(b, prop.uuid(), v_int(666)).unwrap();
+    let (prop, v, perms, is_clear) = tx.resolve_property(b, "test".into()).unwrap();
+    assert_eq!(prop.name(), "test");
+    assert_eq!(v, v_int(666));
+    assert_eq!(perms.owner(), b);
+    assert_eq!(perms.flags(), BitEnum::new_with(PropFlag::Read));
+    assert!(!is_clear);
+
+    assert_eq!(tx.commit(), Ok(CommitResult::Success));
+}
+
+pub fn perform_test_rename_property<F, TX>(begin_tx: F)
+where
+    F: Fn() -> RelationalWorldStateTransaction<TX>,
+    TX: RelationalTransaction<WorldStateTable>,
+{
+    let tx = begin_tx();
+    let a = tx
+        .create_object(
+            None,
+            ObjAttrs::new(NOTHING, NOTHING, NOTHING, BitEnum::new(), "test"),
+        )
+        .unwrap();
+
+    let b = tx
+        .create_object(
+            None,
+            ObjAttrs::new(NOTHING, a, NOTHING, BitEnum::new(), "test2"),
+        )
+        .unwrap();
+
+    let uuid = tx
+        .define_property(
+            a,
+            a,
+            "test".into(),
+            NOTHING,
+            BitEnum::new(),
+            Some(v_str("test_value")),
+        )
+        .unwrap();
+
+    // I can update the name on the parent...
+    tx.update_property_info(a, uuid, None, None, Some("a_new_name".to_string()))
+        .unwrap();
+
+    // And now resolve that new name on the child.
+    let (prop, v, perms, is_clear) = tx.resolve_property(b, "a_new_name".into()).unwrap();
+    assert_eq!(prop.name(), "a_new_name");
+    assert_eq!(v, v_str("test_value"));
+    assert_eq!(perms.owner(), NOTHING);
+    assert!(is_clear);
+
+    // But it's illegal to try to rename it on the child who doesn't define it.
+    assert!(tx
+        .update_property_info(b, uuid, None, None, Some("a_new_name".to_string()))
+        .is_err())
+}
+
+/// Test regression where parent properties were present via `properties()` on children.
+pub fn perform_test_regression_properties<F, TX>(begin_tx: F)
+where
+    F: Fn() -> RelationalWorldStateTransaction<TX>,
+    TX: RelationalTransaction<WorldStateTable>,
+{
+    let tx = begin_tx();
+
+    let a = tx
+        .create_object(
+            None,
+            ObjAttrs::new(NOTHING, NOTHING, NOTHING, BitEnum::new(), "test"),
+        )
+        .unwrap();
+
+    let b = tx
+        .create_object(
+            None,
+            ObjAttrs::new(NOTHING, a, NOTHING, BitEnum::new(), "test2"),
+        )
+        .unwrap();
+
+    // Define 1 property on parent
+    tx.define_property(
+        a,
+        a,
+        "test".into(),
+        NOTHING,
+        BitEnum::new(),
+        Some(v_str("test_value")),
+    )
+    .unwrap();
+    let (prop, v, perms, is_clear) = tx.resolve_property(b, "test".into()).unwrap();
+    assert_eq!(prop.name(), "test");
+    assert_eq!(v, v_str("test_value"));
+    assert_eq!(perms.owner(), NOTHING);
+    assert!(is_clear);
+
+    // And another on child
+    let child_prop = tx
+        .define_property(
+            b,
+            b,
+            "test2".into(),
+            NOTHING,
+            BitEnum::new(),
+            Some(v_str("test_value2")),
+        )
+        .unwrap();
+
+    let props = tx.get_properties(b).unwrap();
+
+    // Our prop should be there
+    assert!(props.find(&child_prop).is_some());
+
+    // Listing the set of properties on the child should include only the child's properties
+    assert_eq!(props.len(), 1);
+}
+
+pub fn perform_test_verb_resolve<F, TX>(begin_tx: F)
+where
+    F: Fn() -> RelationalWorldStateTransaction<TX>,
+    TX: RelationalTransaction<WorldStateTable>,
+{
+    let tx = begin_tx();
+
+    let a = tx
+        .create_object(
+            None,
+            ObjAttrs::new(NOTHING, NOTHING, NOTHING, BitEnum::new(), "test"),
+        )
+        .unwrap();
+
+    tx.add_object_verb(
+        a,
+        a,
+        vec!["test".into()],
+        vec![],
+        BinaryType::LambdaMoo18X,
+        BitEnum::new(),
+        VerbArgsSpec::this_none_this(),
+    )
+    .unwrap();
+
+    assert_eq!(
+        tx.resolve_verb(a, "test".into(), None).unwrap().names(),
+        vec!["test"]
+    );
+
+    assert_eq!(
+        tx.resolve_verb(a, "test".into(), Some(VerbArgsSpec::this_none_this()))
+            .unwrap()
+            .names(),
+        vec!["test"]
+    );
+
+    let v_uuid = tx.resolve_verb(a, "test".into(), None).unwrap().uuid();
+    assert_eq!(tx.get_verb_binary(a, v_uuid).unwrap(), vec![]);
+
+    // Add a second verb with a different name
+    tx.add_object_verb(
+        a,
+        a,
+        vec!["test2".into()],
+        vec![],
+        BinaryType::LambdaMoo18X,
+        BitEnum::new(),
+        VerbArgsSpec::this_none_this(),
+    )
+    .unwrap();
+
+    // Verify we can get it
+    assert_eq!(
+        tx.resolve_verb(a, "test2".into(), None).unwrap().names(),
+        vec!["test2"]
+    );
+    assert_eq!(tx.commit(), Ok(CommitResult::Success));
+
+    // Verify existence in a new transaction.
+    let tx = begin_tx();
+    assert_eq!(
+        tx.resolve_verb(a, "test".into(), None).unwrap().names(),
+        vec!["test"]
+    );
+    assert_eq!(
+        tx.resolve_verb(a, "test2".into(), None).unwrap().names(),
+        vec!["test2"]
+    );
+    assert_eq!(tx.commit(), Ok(CommitResult::Success));
+}
+
+pub fn perform_test_verb_resolve_inherited<F, TX>(begin_tx: F)
+where
+    F: Fn() -> RelationalWorldStateTransaction<TX>,
+    TX: RelationalTransaction<WorldStateTable>,
+{
+    let tx = begin_tx();
+
+    let a = tx
+        .create_object(
+            None,
+            ObjAttrs::new(NOTHING, NOTHING, NOTHING, BitEnum::new(), "test"),
+        )
+        .unwrap();
+
+    let b = tx
+        .create_object(
+            None,
+            ObjAttrs::new(NOTHING, a, NOTHING, BitEnum::new(), "test2"),
+        )
+        .unwrap();
+
+    tx.add_object_verb(
+        a,
+        a,
+        vec!["test".into()],
+        vec![],
+        BinaryType::LambdaMoo18X,
+        BitEnum::new(),
+        VerbArgsSpec::this_none_this(),
+    )
+    .unwrap();
+
+    assert_eq!(
+        tx.resolve_verb(b, "test".into(), None).unwrap().names(),
+        vec!["test"]
+    );
+
+    assert_eq!(
+        tx.resolve_verb(b, "test".into(), Some(VerbArgsSpec::this_none_this()))
+            .unwrap()
+            .names(),
+        vec!["test"]
+    );
+
+    let v_uuid = tx.resolve_verb(b, "test".into(), None).unwrap().uuid();
+    assert_eq!(tx.get_verb_binary(a, v_uuid).unwrap(), vec![]);
+    assert_eq!(tx.commit(), Ok(CommitResult::Success));
+}
+
+pub fn perform_test_verb_resolve_wildcard<F, TX>(begin_tx: F)
+where
+    F: Fn() -> RelationalWorldStateTransaction<TX>,
+    TX: RelationalTransaction<WorldStateTable>,
+{
+    let tx = begin_tx();
+    let a = tx
+        .create_object(
+            None,
+            ObjAttrs::new(NOTHING, NOTHING, NOTHING, BitEnum::new(), "test"),
+        )
+        .unwrap();
+
+    let verb_names = vec!["dname*c", "iname*c"];
+    tx.add_object_verb(
+        a,
+        a,
+        verb_names.iter().map(|s| s.to_string()).collect(),
+        vec![],
+        BinaryType::LambdaMoo18X,
+        BitEnum::new(),
+        VerbArgsSpec::this_none_this(),
+    )
+    .unwrap();
+
+    assert_eq!(
+        tx.resolve_verb(a, "dname".into(), None).unwrap().names(),
+        verb_names
+    );
+
+    assert_eq!(
+        tx.resolve_verb(a, "dnamec".into(), None).unwrap().names(),
+        verb_names
+    );
+
+    assert_eq!(
+        tx.resolve_verb(a, "iname".into(), None).unwrap().names(),
+        verb_names
+    );
+
+    assert_eq!(
+        tx.resolve_verb(a, "inamec".into(), None).unwrap().names(),
+        verb_names
+    );
+    assert_eq!(tx.commit(), Ok(CommitResult::Success));
+}
+
+pub fn perform_reparent_props<F, TX>(begin_tx: F)
+where
+    F: Fn() -> RelationalWorldStateTransaction<TX>,
+    TX: RelationalTransaction<WorldStateTable>,
+{
+    let tx = begin_tx();
+    let a = tx
+        .create_object(
+            None,
+            ObjAttrs::new(NOTHING, NOTHING, NOTHING, BitEnum::new(), "test"),
+        )
+        .unwrap();
+    let b = tx
+        .create_object(
+            None,
+            ObjAttrs::new(NOTHING, a, NOTHING, BitEnum::new(), "test2"),
+        )
+        .unwrap();
+    let c = tx
+        .create_object(
+            None,
+            ObjAttrs::new(NOTHING, b, NOTHING, BitEnum::new(), "test3"),
+        )
+        .unwrap();
+
+    // Add a property on A
+    tx.define_property(
+        a,
+        a,
+        "test".into(),
+        NOTHING,
+        BitEnum::new(),
+        Some(v_str("test_value")),
+    )
+    .unwrap();
+
+    // Verify it's on B & C
+    let (prop, v, perms, is_clear) = tx.resolve_property(b, "test".into()).unwrap();
+    assert_eq!(prop.name(), "test");
+    assert_eq!(v, v_str("test_value"));
+    assert_eq!(perms.owner(), NOTHING);
+    assert!(is_clear);
+
+    let (prop, v, perms, is_clear) = tx.resolve_property(c, "test".into()).unwrap();
+    assert_eq!(prop.name(), "test");
+    assert_eq!(v, v_str("test_value"));
+    assert_eq!(perms.owner(), NOTHING);
+    assert!(is_clear);
+
+    // Now make a new root and reparent B to it
+    let d = tx
+        .create_object(
+            None,
+            ObjAttrs::new(NOTHING, NOTHING, NOTHING, BitEnum::new(), "test4"),
+        )
+        .unwrap();
+
+    tx.set_object_parent(b, d).unwrap();
+
+    // Verify the property is no longer on B
+    let result = tx.resolve_property(b, "test".into());
+    assert_eq!(
+        result.err().unwrap(),
+        WorldStateError::PropertyNotFound(b, "test".into())
+    );
+
+    // Or C.
+    let result = tx.resolve_property(c, "test".into());
+    assert_eq!(
+        result.err().unwrap(),
+        WorldStateError::PropertyNotFound(c, "test".into())
+    );
+
+    // Now add new property on D
+    tx.define_property(
+        d,
+        d,
+        "test2".into(),
+        NOTHING,
+        BitEnum::new(),
+        Some(v_str("test_value2")),
+    )
+    .unwrap();
+
+    // Verify it's on B and C
+    let (prop, v, perms, is_clear) = tx.resolve_property(b, "test2".into()).unwrap();
+    assert_eq!(prop.name(), "test2");
+    assert_eq!(v, v_str("test_value2"));
+    assert_eq!(perms.owner(), NOTHING);
+    assert!(is_clear);
+
+    let (prop, v, perms, is_clear) = tx.resolve_property(c, "test2".into()).unwrap();
+    assert_eq!(prop.name(), "test2");
+    assert_eq!(v, v_str("test_value2"));
+    assert_eq!(perms.owner(), NOTHING);
+    assert!(is_clear);
+
+    // And now reparent C to A again, and verify it's back to having the first property.
+    tx.set_object_parent(c, a).unwrap();
+    let (prop, v, perms, is_clear) = tx.resolve_property(c, "test".into()).unwrap();
+    assert_eq!(prop.name(), "test");
+    assert_eq!(v, v_str("test_value"));
+    assert_eq!(perms.owner(), NOTHING);
+    assert!(is_clear);
+}
